+++ src/hiero_sdk_python/__init__.py
@@ -5,6 +5,7 @@ from .client.network import Network
 # Account
 from .account.account_id import AccountId
 from .account.account_create_transaction import AccountCreateTransaction
+from .account.account_update_transaction import AccountUpdateTransaction
 from .account.account_info import AccountInfo
 
 # Crypto
@@ -27,6 +28,10 @@ from .tokens.token_burn_transaction import TokenBurnTransaction
 from .tokens.token_grant_kyc_transaction import TokenGrantKycTransaction
 from .tokens.token_revoke_kyc_transaction import TokenRevokeKycTransaction
 from .tokens.token_update_transaction import TokenUpdateTransaction
+from .tokens.token_airdrop_transaction import TokenAirdropTransaction
+from .tokens.token_cancel_airdrop_transaction import TokenCancelAirdropTransaction
+from .tokens.pending_airdrop_id import PendingAirdropId
+from .tokens.pending_airdrop_record import PendingAirdropRecord
 from .tokens.token_id import TokenId
 from .tokens.token_type import TokenType
 from .tokens.supply_type import SupplyType
@@ -81,8 +86,23 @@ from .logger.log_level import LogLevel
 
 # File
 from .file.file_create_transaction import FileCreateTransaction
+from .file.file_append_transaction import FileAppendTransaction
 from .file.file_info_query import FileInfoQuery
 from .file.file_info import FileInfo
+from .file.file_contents_query import FileContentsQuery
+from .file.file_update_transaction import FileUpdateTransaction
+from .file.file_delete_transaction import FileDeleteTransaction
+
+# Contract
+from .contract.contract_create_transaction import ContractCreateTransaction
+from .contract.contract_call_query import ContractCallQuery
+from .contract.contract_info_query import ContractInfoQuery
+from .contract.contract_bytecode_query import ContractBytecodeQuery
+from .contract.contract_execute_transaction import ContractExecuteTransaction
+from .contract.contract_function_parameters import ContractFunctionParameters
+from .contract.contract_function_result import ContractFunctionResult
+from .contract.contract_info import ContractInfo
+from .contract.contract_update_transaction import ContractUpdateTransaction
 
 __all__ = [
     # Client
@@ -92,6 +112,7 @@ __all__ = [
     # Account
     "AccountId",
     "AccountCreateTransaction",
+    "AccountUpdateTransaction",
     "AccountInfo",
 
     # Crypto
@@ -118,6 +139,10 @@ __all__ = [
     "TokenGrantKycTransaction",
     "TokenRelationship",
     "TokenUpdateTransaction",
+    "TokenAirdropTransaction",
+    "TokenCancelAirdropTransaction",
+    "PendingAirdropId",
+    "PendingAirdropRecord",
     "TokenType",
     "SupplyType",
 
@@ -147,11 +172,11 @@ __all__ = [
     "TokenNftInfoQuery",
     "TokenInfoQuery",
     "AccountInfoQuery",
-    
+
     # Address book
     "Endpoint",
     "NodeAddress",
-    
+
     # Logger
     "Logger",
     "LogLevel",
@@ -161,9 +186,24 @@ __all__ = [
     "ResponseCode",
     "Timestamp",
     "Duration",
-    
+
     # File
     "FileCreateTransaction",
+    "FileAppendTransaction",
     "FileInfoQuery",
     "FileInfo",
+    "FileContentsQuery",
+    "FileUpdateTransaction",
+    "FileDeleteTransaction",
+
+    # Contract
+    "ContractCreateTransaction",
+    "ContractCallQuery",
+    "ContractInfoQuery",
+    "ContractBytecodeQuery",
+    "ContractExecuteTransaction",
+    "ContractFunctionParameters",
+    "ContractFunctionResult",
+    "ContractInfo",
+    "ContractUpdateTransaction",
 ]