+++ src/hiero_sdk_python/consensus/topic_delete_transaction.py
@@ -1,9 +1,28 @@
+"""
+This module provides the `TopicDeleteTransaction` class for deleting consensus topics
+on the Hedera network using the Hiero SDK.
+
+It handles setting the target topic ID, building the protobuf transaction body, and
+defining the execution method required to perform the deletion transaction.
+"""
+
+
 from hiero_sdk_python.transaction.transaction import Transaction
-from hiero_sdk_python.hapi.services import consensus_delete_topic_pb2, transaction_body_pb2, basic_types_pb2
+from hiero_sdk_python.hapi.services import (
+    consensus_delete_topic_pb2,
+    transaction_body_pb2,
+    basic_types_pb2
+)
 from hiero_sdk_python.channels import _Channel
 from hiero_sdk_python.executable import _Method
 
+
 class TopicDeleteTransaction(Transaction):
+    """
+        Represents a transaction to delete an existing topic in the Hedera
+        Consensus Service (HCS).
+
+    """
     def __init__(self, topic_id: basic_types_pb2.TopicID = None):
         super().__init__()
         self.topic_id: basic_types_pb2.TopicID = topic_id
@@ -35,9 +54,9 @@ class TopicDeleteTransaction(Transaction):
         """
         if self.topic_id is None:
             raise ValueError("Missing required fields: topic_id")
-    
         transaction_body: transaction_body_pb2.TransactionBody = self.build_base_transaction_body()
-        transaction_body.consensusDeleteTopic.CopyFrom(consensus_delete_topic_pb2.ConsensusDeleteTopicTransactionBody(
+        transaction_body.consensusDeleteTopic.CopyFrom(
+            consensus_delete_topic_pb2.ConsensusDeleteTopicTransactionBody(
             topicID=self.topic_id._to_proto()
         ))