+++ tests/integration/topic_info_query_e2e_test.py
@@ -29,8 +29,8 @@ def test_integration_topic_info_query_can_execute():
         
         assert topic_info.memo == "Topic for info query"
         assert topic_info.sequence_number == 0
-        assert env.client.operator.private_key.public_key().to_string() == topic_info.admin_key.ed25519.hex()
-        
+        assert env.client.operator_private_key.public_key()._to_proto() == topic_info.admin_key
+
         delete_transaction = TopicDeleteTransaction(topic_id=topic_id)
         delete_transaction.freeze_with(env.client)
         delete_receipt = delete_transaction.execute(env.client)